Fix compile error
