+++ core/model/src/main/kotlin/com/stash/core/model/QobuzDiscoveryStatus.kt
@@ -1,3 +1,3 @@
-package com.stash.core.data.discovery
+package com.stash.core.model.discovery
 
 enum class QobuzDiscoveryStatus { OK, NO_TOKEN, NO_INTERNET }
